+++ contracts/token/src/admin.rs
@@ -8,11 +8,13 @@ pub fn has_administrator(e: &Env) -> bool {
     e.storage().instance().has(&key)
 }
 
+#[allow(dead_code)]
 pub fn read_administrator(e: &Env) -> Address {
     let key = DataKey::Admin;
     e.storage().instance().get(&key).unwrap()
 }
 
+#[allow(dead_code)]
 pub fn write_administrator(e: &Env, id: &Address) {
     let key = DataKey::Admin;
     e.storage().instance().set(&key, id);