+++ contracts/token/src/metadata.rs
@@ -16,7 +16,6 @@ pub fn read_symbol(e: &Env) -> String {
     util.metadata().get_metadata().symbol
 }
 
-#[allow(dead_code)]
 pub fn write_metadata(e: &Env, metadata: TokenMetadata) {
     let util = TokenUtils::new(e);
     util.metadata().set_metadata(&metadata);